Extract VeeValidate config
