+++ src/stories/AeModal.stories.js
@@ -1,12 +1,11 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import Vue from 'vue';
 import Vuex from 'vuex';
-import VeeValidate, { Validator } from 'vee-validate';
-import BigNumber from 'bignumber.js';
 import { storiesOf } from '@storybook/vue';
 import { action } from '@storybook/addon-actions';
 import '@aeternity/aepp-components-3/dist/aepp.components.css';
 import '@aeternity/aepp-components-3/dist/aepp.fonts.css';
+import VeeValidate from '../lib/veeValidatePlugin';
 import AeButton from '../components/AeButton.vue';
 import MigratedBalanceModal from '../components/MigratedBalanceModal.vue';
 import LedgerModal from '../components/desktop/LedgerModal.vue';
@@ -16,23 +15,7 @@ import TransactionFeeModal from '../components/desktop/TransactionFeeModal.vue';
 import { account } from './mock-data';
 
 Vue.use(Vuex);
-Validator.extend('min_value_exclusive', (value, [min]) => BigNumber(value).isGreaterThan(min));
-Validator.extend('min_value', (value, [max]) => BigNumber(value).isGreaterThanOrEqualTo(max));
-Validator.extend('max_value', (value, [max]) => BigNumber(value).isLessThanOrEqualTo(max));
-
-Vue.use(VeeValidate, {
-  dictionary: {
-    en: {
-      messages: {
-        required: 'This field is required',
-        min: (field, [length]) => `This field must be at least ${length} characters`,
-        min_value: (field, [min]) => `This field must be ${min} or more`,
-        min_value_exclusive: (field, [min]) => `This field must be more than ${min}`,
-        max_value: (field, [max]) => `This field must be ${max} or less`,
-      },
-    },
-  },
-});
+Vue.use(VeeValidate);
 
 storiesOf('AeModal', module)
   .add('MigratedBalanceModal', () => ({
